refactor: add api operation on users

+++ backend/src/users/users.controller.ts
@@ -6,8 +6,14 @@ import {
   NotFoundException,
   Param,
   Patch,
+  Req,
+  UnauthorizedException,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
+import { ApiOperation, ApiResponse } from '@nestjs/swagger';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { JwtPayload } from 'src/auth/entities/auth.entity';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserInterceptor } from './interceptors/user.interceptor';
 import { UsersService } from './users.service';
@@ -18,11 +24,20 @@ export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Get()
+  @UseGuards(AuthGuard)
+  @ApiOperation({ summary: 'Get all users' })
+  @ApiResponse({ status: 200, description: 'Users retrieved successfully' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
   findAll() {
     return this.usersService.findAll();
   }
 
   @Get(':id')
+  @UseGuards(AuthGuard)
+  @ApiOperation({ summary: 'Get a user by ID' })
+  @ApiResponse({ status: 200, description: 'User retrieved successfully' })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @ApiResponse({ status: 404, description: 'User not found' })
   async findOne(@Param('id') id: string) {
     const user = await this.usersService.findOne(id);
     if (!user) throw new NotFoundException('User not found');
@@ -30,12 +45,36 @@ export class UsersController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @UseGuards(AuthGuard)
+  @ApiOperation({ summary: 'Update a user' })
+  @ApiResponse({ status: 200, description: 'User updated successfully' })
+  @ApiResponse({
+    status: 403,
+    description: 'Forbidden: Not the owner of this account',
+  })
+  update(
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Req() req: Request & { user: JwtPayload },
+  ) {
+    if (id !== req.user.id) {
+      throw new UnauthorizedException('You can only update your own account');
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard)
+  @ApiOperation({ summary: 'Delete a user' })
+  @ApiResponse({ status: 200, description: 'User deleted successfully' })
+  @ApiResponse({
+    status: 403,
+    description: 'Forbidden: Not the owner of this account',
+  })
+  remove(@Param('id') id: string, @Req() req: Request & { user: JwtPayload }) {
+    if (id !== req.user.id) {
+      throw new UnauthorizedException('You can only delete your own account');
+    }
     return this.usersService.remove(id);
   }
 }
